I find the code generally over-commented
This is fine really - comments are supposed to explain what
you're doing if it's non-obvious from the code, and what's
obvious is of course subjective. Generally though explaining
what a function does should happen in a docstring, or in the
name of a function. If you're tempted to use a comment to describe
what lines of code do, consider making that line or two their own
function and naming the function like you were going to write the
comment

Comments are scary if they just describe the code because one might
change the code without changing the comment.